debian/extra/kernel-install.d/85-initrd.install: Remove an unnecessary variable.
When removing, kernel-install calls its plugins with only three arguments [0]
(in contrast to adding where it's four arguments). Since we are running with
"set -u", trying to expand the fourth argument causes the script to fail.
However, as the $KERNEL_IMAGE variable is never used, we can simply remove it.

[0] https://github.com/systemd/systemd/blob/v232/src/kernel-install/kernel-install#L140

Closes: #845977