Simplify version checks for save_is_enabled
Stop testing for specific versions. This is no longer necessary as even
the versions in jessie satisfy those requirements.

Also, run the code on upgrades only. Trying to do that on new installs
(even if there is state from a removed but not purged systemd package)
is rather pointless, as systemctl will not be available at this point.