Avoid strict DM API versioning
Compiling against the dm-ioctl.h header as provided by the Linux kernel
will embed the DM interface version number. Running an older kernel can
lead to errors on shutdown when trying to detach DM devices.
As a workaround, build against a local copy of dm-ioctl.h based on 3.13,
which is the minimum required version to support DM_DEFERRED_REMOVE.

Closes: #856337