ubuntu: disable dnssec on any ubuntu releases
The internet is broken, and debugging the internet with Ubuntu is not
helpful. Too many websites are incorrectly signed with dnssec, and there are
many outstanding bugs upstream and newly reported in Ubuntu as soon as artful
landed with dnssec re-enabled. Ubuntu devel releases are used on day to day
basis and are not experimental enough to break developers'
networking. Re-enabling dnssec should only be considered once existing upstream
and launchpad dnssec bugs are resolved.

LP: #1690605
Gbp-Dch: Short