Bump debhelper compatibility level to 10
Drop --with autoreconf and --parallel as those are now enabled by
default.

The systemd sequence is now also enabled by default. We don't strictly
need the additional complexity that comes with init-system-helpers, as
we can just rely on systemctl being available. So use --without systemd
for the time being.

Gbp-Dch: Short