udev README.Debian: Include initrd rebuild and some clarifications in migration
While initrd update is already being mentioned in the introductory
section, it is easy to miss when going through the migration steps, so
explicitly mention it again. Also add a warning about keeping a fallback
on misconfigurations, and the possibility to migrate one interface at a
time.

Thanks to Karl O. Pinc for the suggestions!

Closes: #881769