Drop seccomp system call filter for udev
The seccomp based system call whitelist requires at least systemd 239 to
be the active init and during a dist-upgrade we can't guarantee that
systemd has been fully configured before udev is restarted.
The versioned systemd Breaks that was added to udev for #902185 didn't
really fix this issue, so revert that change again.

Closes: #903224