Run "meson test" instead of "ninja test"
Upstream developers of meson recommend to run it in this way, because "ninja
test" just calls "meson test", and by using meson directly and using extra
command line arguments it is possible to control aspects of how the tests are
run.

For example, it is possible to increase the timeouts depending on the
architecture (for slow buildds), which otherwise it is not possible without
resorting to patch upstream's build files.

For more details, see:
 - https://bugs.debian.org/906429
 - https://github.com/mesonbuild/meson/issues/2037