debian/tests/boot-smoke: Create "render" system group
When running tests for upstream PRs, this test often fails with

    checking for connection timeouts
    systemd-udevd[1228]: Failed to resolve group 'render': Connection timed out

Which is not the kind of timeout the test is looking for. Create the
group in the test to avoid this.

We explicitly don't create the group in systemd.postinst as we revert
the patch that introduces the group into the udev rules.