d/t/boot-and-services: check for any kernel message, not just first kernel message
The test currently looks for the first(ish) kernel log message, which may
not be present.  If it's not, the test case fails.

This isn't the fault of systemd/journald, the problem is the kernel filled
up its klog buffer before we started journald to read them.  This can be
caused by a too-small kernel klog buffer, or could be caused by a large
number of kernel boot-time messages.  More details are in this Ubuntu bug:
https://bugs.launchpad.net/ubuntu/+source/systemd/+bug/1830479

While catching all the kernel log messages since boot is important, this
test case is testing systemd, not the kernel, and should not fail if
the kernel's log buffer is filling during boot.

This problem has caused upstream systemd to disable Ubuntu CI on arm64
for many months:
https://github.com/systemd/systemd/issues/11104

Closes: #929730