d/t/upstream: add TEST-30, TEST-34 to blacklist
Unfortunately, these tests are failing on i386 builds, very
intermittently; this causes headaches for upstream systemd
maintainers and submitters, when the Ubuntu CI fails, because
the failure isn't related to the PRs the tests are running for,
and determining that from the Ubuntu CI log file and other artifacts
is very time-consuming.

This just blacklists the tests for now, until we can figure out why
they are failing and fix them.

TEST-30 is discussed:
https://github.com/systemd/systemd/issues/12268

TEST-34 is discussed:
https://github.com/systemd/systemd/issues/12932