d/e/checkout-upstream: create git commits for each change
Create a commit with any debian/ changes, commit with all upstream code
changes, and a final commit with the updated changelog entry.

For normal operation, where this script is being called from upstream
autopkgtest tests, there is no need for creating git commits with the
changes made by this script; however, creating the commits won't (shouldn't)
break anything in the normal upstream test process.

Where creating git commits with this script's changes is useful is for a
local git repository used for development - specifically, to maintain
a git repo that includes the upstream systemd source code, but also the
debian/ code (including tests) so that the package can be built in a ppa
and tested using autopkgtest, without needing to use github to perform the
testing.