d/e/checkout-upstream: add UPSTREAM_KEEP_CHANGELOG param
It can be useful to skip creating the new changelog entry, if
the caller only wants to pull the latest upstream code into a
git repo; this adds a env var that allows doing that.

By default, without the var set, the changelog entry is created,
so existing behavior is unchanged.