d/t/storage: wait for service to start, only stop if active
This removes the systemctl start, as the service should already be
running (or starting, or failed).  Instead, just wait for the service
to actually reach 'active' or 'failed' state.  Then, only stop the
service if its state is 'active'.

Previous attempts to fix hung or flaky test runs included adding a start
before the stop, but the start-the-stop calls can lead to cancelled jobs
instead of a properly stopped service, which is again flaky.