config: tpm-pcrindex=12, on Ubuntu only.
In Ubuntu, we are trying to integrate measured boot with either bootctl or
grub, or sometimes both using chainloading. In those cases we may want to seal
against grub's PCRs, bootctl stub PCRs, or both at the same time. To achieve
that, we wouldn't like to decouple grub's & bootctl's PCR and make the latter
one use PCR #12.