Follow symlinks when finding link files to copy into initramfs
If the /{etc,lib}/systemd/network directory itself is a symlink, the find
command will not actually find any of the files in the dir it links to.
Use the find -L param to follow symlinks.

Note that the -L does not need to be provided to the cp command, as when
using only the -p parameter symlinks are followed by default.  Also,
the [ -d ] test follows symlinks by default, and does not need changing.

LP: #1868892