root-unittests: remove any corrupt system journal files before testing
Corrupt journal files may cause some tests to fail, most often
test-journal-flush.

This may be fixed better in code upstream, but for SRU purposes we
do not need to test corrupt journal files, as (in autopkgtest context)
they simply indicate a problem with the test instance.

For the specific LP bug, the corruption appears to be caused on the arm64
instances due to a problem rebooting the instance, that corrupts the journal
file(s) on reboot.

LP: #1881947