Add systemd-boot-efi multiarch package
Allows EFI binaries for different architectures to be co-installed.
Useful when the EFI has a different architecture, or to manipulate
images. The userspace tooling doesn't need to match the EFI binaries.

Also allows to reduce the number of packages and dependencies
needed when i386 is not a full architecture, but a subset for libraries
and for EFI support.