Use a single NEWS file shipped in the main systemd package
There is a long standing issue in apt-listchanges which fails to show
NEWS entries if they come from different binary packages that are built
from the same source package.

See https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=977422

It is questionable if this bug is actually fixable, so install a single
NEWS file instead.

Using debian/NEWS would copy that file into every binary package. To
avoid that duplication, we use debian/systemd.NEWS instead, so it is only
shipped in the main systemd package.

This ensures that NEWS entries are reliably shown on upgrades of the
systemd package even if it means that irrelevant NEWS entries are shown
as well, say for users that do not have systemd-journal-remote
installed.

Gbp-Dch: Short