Drop Add-support-for-TuxOnIce-hibernation.patch
This patch adds support for TuxOnIce hibernation, which requires an
out-of-tree kernel patch to be operable. Given that this functionality
is not appropriate for upstream systemd, the kernel never accepted
TuxOnIce, and the project appears to be largely unmaintained, it does
not seem appropriate to carry this patch in Ubuntu.

See [1] for original Debian discussion of this patch.

[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=746463