test: skip some tests when machine-id is not initialized
Upstream tests can currently handle when /etc/machine-id does not exist,
but it does not handle the cases where it is unitialized. In Ubuntu,
this only affects i386 builds for some reason, but would affect any
chroot build where /etc/machine-id is empty.

At the time of writing, it is unclear if it would be best to fix this
upstream, or to remove the Ubuntu hack from live-build that truncates
/etc/machine-id instead of just removing it.

Gbp-Dch: Short