debian/tests: stop testing in privileged LXD containers
LoadCredential is broken in privileged LXD containers, which is expected
due to the apparmor policies in place for privileged containers[1].
Currently we have a work around for this in tests-in-lxd to override
systemd-sysusers.service so that it does not load credentials. Now, the
systemd-tmpfile-* services also attempt to load credentials, which means
these services fail in privileged LXD containers and lead to various
other problems.

Rather than continuing to work around the quirks of privileged LXD
containers, just stop testing them. See [3] for more justification.

[1] https://bugs.launchpad.net/ubuntu/+source/lxd/+bug/1950787/comments/3
[2] https://launchpad.net/bugs/1950787
[3] https://bugs.launchpad.net/ubuntu/+source/lxd/+bug/1950787/comments/4

Gbp-Dch: Short