60-keyboard: Add Pang12 config
Fn+F6 has a disable touchpad icon on it, but the mapping was
incorrect. This fixes that.

As well Fn+End isn't labeled but does have a keyboard value.
Without changing that gnome interprets the default mapping as a
"Audio Next" key, which when no music is playing gives a default
"Action Unavailable" icon in Gnome. This remaps that as a
programable key.