udev: gracefully handle rename failures (LP: #2002445)
These patches are follow-on to commit 1b91e083aabb59213d62c26d993f9be1e0d1dff7
("udev: avoid NIC renaming race with kernel (LP: #2002445)"). With the
previous patches, the error from the kernel is handled gracefully and
the would-have-been new name is kept as an alternative name. However,
udev did not previously have logic to undo the other changes made to
device properties.

These patches complete the handling of rename failues by adding logic to
restore syspath and properties on rename failure, including but not
limited to the case where the kernel will not allow the device rename.

This consists of the following upstream commits:

 * core/device: ignore failed uevents
 * sd-device: introduce device_get_property_int()
 * udev: restore syspath and properties on failure
 * sd-device: make device_set_syspath() clear sysname and sysnum

Gbp-Dch: Short