debian/tests/boot-and-services: drop test_no_failed
When this test fails, it is almost always because of units from outside
of src:systemd, and the resolution is usually to just ignore the
problematic unit. There are currently plenty of exceptions, and this
test has become a bit of a nuisance.

Since we cannot control the quality of external units, just drop this
test to avoid unnecessary blockers in the future.

Gbp-Dch: Short