pstore: only try to load efi_pstore module (LP: #2023462)
The original purpose of this commit was to make sure that efi_pstore was
around, but then the author tried to be all-inclusive and load all
pstore providers. This caused annoyance, particularly around mtdpstore,
so that was dropped (LP: #1981622). But even then, we only dropped
mtdpstore because that was the pain point. However, in upstream we
ended up with ONLY efi_pstore (https://github.com/systemd/systemd/commit/8b8bd621e1),
and it would have been best if we just did that in the SRU that removed
mtdpstore too.

Do that now, since leaving the other modules around can cause pstore
provider modules to be loaded needlessly.

Gbp-Dch: Short