Drop debian/UBUNTU-wait-online-exit-if-no-links-are-managed.patch (LP: #1982218)
The original intent of this patch is to exit early when no devices are
managed by networkd, and was developed against bionic. Since then,
wait-online gain an --any flag, and this code changed the context of
this patch. Namely, the logic added by the patch will not be reached
unless the --any flag is passed, which means the patch no longer works
as expected (systemd-networkd-wait-online.service does not use the
--any flag).

Further, if the --any flag is used, this patch will make wait-online
exit while a link is still in "configuring" state, which can mean that
units which depend on networkd-online.target may not actually have a
functioning link yet.

Because this patch is not serving it's original purpose, and is creating
a new issue, it seems the best thing to do is just drop the patch.

Gbp-Dch: Short