duplicate udev trigger interest for /usr-merge
As packages move their files from / to /usr, dpkg will cease activating
udev's trigger interest for hwdb.d files. In DEP17, this is documented
as P2. Since there are only two packages in Debian affected by this
issue, we opt for an individual mitigation and explicitly duplicate the
trigger interest to both locations (M12). This is safe regardless of the
other chosen mitigations as trigger activations are accumulated.