debian/rules: build with -ffat-lto-objects (LP: #2031416)
Building with -flto=auto, but without -ffat-lto-objects, causes static
archives to end up without any code sections. This is needed to ship a
usable libsystemd.a in libsystemd-dev.

On Ubuntu, this flag is on by default on most arches, but not on armhf.

See https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=977596.

Gbp-Dch: Short