Move kernel-install initrd script to earlier prefix
From systemd v255 onwards, the 60-ukify.install script automatically
uses initrd files that are present in the staging area. Move the initrd
script to run earlier, so that ukify can use the initrd.

The only third-party kernel-install script present in the Debian archive
is 40-dkms.install, so it is not impacted by this change. Both dracut
and mkinitcpio upstream also ship their script with priority 50.