debian/systemd.install: exclude files that are not built for i386
Almost all of these are related to the tpm2 tooling, except for bsod,
which is not built on i386 because we don't build with libqrencode on
i386.

These would not have been shipped before, but since systemd.install is
more explicit now, we need to explicitly exclude these on i386.

Gbp-Dch: Short