test: skip test-execute in arhmf LXC containers
This has the effect of skipping test-execute on armhf autopkgtest. Many
of the failures we see in LXC are related to AppArmor and kernel
differences, and currently the host for armhf autopkgtest is Focal
running Linux 5.4. Thus, we may get quite different results in
autopkgtest infrastructure than if we ran (a) native armhf host, or (b)
armhf container on a newer kernel (like 6.6 in Noble).

We get better LXC coverage by tests-in-lxd on other arches, and we have
a history already of just skipping things in test-execute on LXC
wholesale when things fail on armhf. Instead of continuing that, just
skip test-execute entirely on armhf in LXC.

Gbp-Dch: Short