debian/rules: don't override default telinit-path
The 'telinit-path' meson option is to be used to point to a non-systemd
telinit binary path. When systemctl is invoked as 'telinit', it will exec
this binary when it detects the current environment is not running under
systemd. If 'telinit-path' points back to 'systemctl', that would result
in an infinite re-execve() loop.

Closes: #1063147