Don't put debian revision in version tag when building for upstream
When building for upstream, the debian revision doesn't really have
any meaning. However, it is still very useful to make sure that every
package built for upstream is newer than the previous one without
changing the upstream version by putting the date in the revision.

To make this work, when building for upstream, don't put the debian
revision in the version tag. Without this change, if we put the date
in the debian revision, the version tag would change each build and
cause a bunch of meson targets to be rebuilt each time, making
incremental builds unnecessarily slow.