Drop debian/Ubuntu-UseDomains-by-default.patch
Overriding a default option for all networks in this way is confusing
and opaque. Most users would find themselves reading upstream systemd
documentation about these settings and their defaults, rather than the
Ubuntu hosted or locally installed manpages.

The patch does not link to a bug to give greater motivation than that
stated in the patch description, so it is unclear how relevant this
still is to its original intention. And, this patch does not apply
anymore due to upstream changes in the implementation of UseDomains=, so
it is a good time to just drop it.

Gbp-Dch: Short