debian/systemd-resolved.postinst: ignore cp failure (LP: #2047975)
In come cases, copying /etc/resolv.conf to /run/systemd/resolve/stub-resolv.conf
will fail, despite the checks that happen beforehand. In particular,
this can happen if a user disabled the stub-resolver, and in doing so,
made /etc/resolv.conf a symlink to /run/systemd/resolve/resolv.conf.
This is unnecessary because systemd-resolved will make stub-resolv.conf
a symlink to resolv.conf if DNSStubListener=no. In these cases, it is
safe to just ignore the cp because it is unnecssary to begin with.

Gbp-Dch: Short