Drop debian/UBUNTU-Don-t-override-Ubuntu-s-default-sysctl-values-LP-1962038.patch
We do not actually ship sysctld.d/50-default.conf in Debian or Ubuntu.
When this patch was written, we did, but that change was lost in a merge
from Debian. The preferred package to set sysctl defaults is procps
anyways, so rather than restoring sysctl.d/50-default.conf to make this
patch relevant, drop the patch.