debian/tests/tests-in-lxd: set security.nesting true for LXD containers
Drop the existing check for skipping the test. A better solution is to
set security.nesting true, which allows more mounts to be done within
the container. While we are not running nested containers per se, many
of the systemd sandboxing settings make some units effectively
containerized via the heavy use of namespaces.

This change is safe because we are using unprivileged containers only.
This change would not be safe with privileged containers.

Gbp-Dch: Short