d/t/boot-and-services: tweak test_rsyslog regexes
This test is currently failing on infra, but not locally, and the cause
is unclear.

The purpose of this test is to be reasonably sure that journal messages
are forwarded to rsyslog. We can check for "init" messages without
looking for graphical.target specifically, can look for services beside
NetworkManager.

Loosen the "init" message regex, and look for systemd-resolved instead
of NetworkManager. This appears more reliable on the infra, maybe
because they are logged about earlier.

Gbp-Dch: Short