d/rules: set SYSTEMD_IN_CHROOT=1 for build tests
For currently unknown reasons, test-reread-partition-table runs at build
time on the build schroots, despite having a running_in_chroot() check
to skip the test in chroots. However, when this test is run manually in
the same environment later, it is correctly skipped.

Manual inspection found that running_in_chroot() is failing with EACCES
during the real build, so the test just continues. The current hypothesis
is that one of the earlier tests is breaking the environment.

For now, take advantage of the SYSTEMD_IN_CHROOT environment variable to
assert that all build time tests should be assumed to be running in a
chroot.

Gbp-Dch: Short