Mirror dmi_arches from meson.build into debian/udev.install
The relevant files are installed for any architecture matched by by
the dmi_arches variable in the upstream meson.build. It does not care
about the libc or kernel, so we start using any patterns here. Notably,
mips was the only architecture missing in the udev.rules. x32 is implied
by any-amd64 and loong64 was duplicated. armel and armhf are both
matches by any-arm as the suffix is part of the libc ABI earlier in the
tuple.