d/t/tests-in-lxd: re-construct --pin-packages arguments for autopkgtest
When --pin-packages is passed to autopkgtest, and --disable-apt-fallback
is NOT passed (default), fallback logic will be invoked if an upgrade
calculation cannot be calculated during testbed setup. This can happen
e.g. testing with a package trigger from experimental.

Unfortunately, in tests-in-lxd, even though we copy in the exact
/etc/apt configs, the fallback logic is not triggered because we didn't
use --pin-packages. This means that it's possible for tests-in-lxd to
fail due to the upgrade issue, even though the same tests in the normal
testbed would succeed after the fallback.

Add a hack to re-construct the --pin-packages arguments based on the
content of /etc/apt/preferences.d/autopkgtest-*.pref.

Gbp-Dch: Short